Moved getting datapoints to abstract class
